+++ cms-20240330/alibabacloud_cms20240330/models/_get_prometheus_instance_request.py
@@ -10,7 +10,7 @@ class GetPrometheusInstanceRequest(DaraModel):
         aliyun_lang: str = None,
         resource_group_id: str = None,
     ):
-        # The language. Default value: zh. Valid values: zh, en.
+        # The language. Default value: zh. Valid values: zh and en.
         self.aliyun_lang = aliyun_lang
         # The resource group ID.
         self.resource_group_id = resource_group_id